Added 4x20 lcdState_handler()
New architecture removes blocking delay() functions.
