+++ Firmware/MVP/libcm.h
@@ -29,7 +29,8 @@
   #include "LiDisplay.h"
   #include "adc.h"
   #include "vPackSpoof.h"
-  #include "lcd.h"
+  #include "lcdState.h"
+  #include "lcdTransmit.h"
   #include "gridCharger.h"
   #include "LTC68042configure.h"
   #include "LTC68042cell.h"
